삭제 테스ㅡㅌ

+++ frontend/src/components/Article.jsx
@@ -7,12 +7,16 @@ export default function Article({ article, isDetail = false, onDelete }) {
   const Navigate = useNavigate();
 
   async function deleteArticle(id) {
-    try {
-      const response = await articlesApi.deleteArticle(id);
-      console.log("삭제 성공:", response);
-      onDelete(id); // 부모 컴포넌트에서 삭제 후 상태 업데이트
-    } catch (error) {
-      console.error("삭제 실패:", error);
+    if (confirm("정말로 삭세하시겠습니까?")) {
+      try {
+        const response = await articlesApi.deleteArticle(id);
+        console.log("삭제 성공:", response);
+        onDelete(id); // 부모 컴포넌트에서 삭제 후 상태 업데이트
+      } catch (error) {
+        console.error("삭제 실패:", error);
+      }
+    } else {
+      return false;
     }
   }
